requirements.txt
python-dotenv
langchain-openai
langchain-community
langchainhub
flask